Restructure builders as independent submodules
