+++ rplugin/python3/buildit/builders/cargo.py
@@ -1,0 +1,10 @@
+''' Builder definitions for Rust Cargo '''
+
+cargo = {
+    'sig': r'Cargo\.toml|Cargo\.lock',
+    'cmd': 'cargo build',
+    'func': None,
+    'ft': 'rust',
+    'subdir': None,
+    'shell': False
+}
