+++ rplugin/python3/buildit/builders/make.py
@@ -1,0 +1,10 @@
+''' Builder definitions for GNU Make '''
+
+make = {
+    'sig': r'Makefile',
+    'cmd': 'make',
+    'func': None,
+    'ft': None,
+    'subdir': None,
+    'shell': False
+}
