+++ rplugin/python3/buildit/builders/stack.py
@@ -1,0 +1,10 @@
+''' Builder definitions for Haskell Stack '''
+
+stack = {
+    'sig': r'stack\.yaml',
+    'cmd': 'stack build',
+    'func': None,
+    'ft': 'haskell',
+    'subdir': None,
+    'shell': False
+}
